+++ src/index.ts
@@ -1,6 +1,4 @@
-export { createLayer } from './api/create-layer'
-export { defineLayer } from './api/define-layer'
-export { LayerTemplate } from './api/layer-template'
+export * from './api'
 export type {
   LayerAdapter,
   LayerConfigFragment,